Refactor: Reorganize inbound order handler into proper package structure
