+++ internal/application/di/inbound_order_dependencies.go
@@ -1,17 +1,17 @@
 package di
 
 import (
-	"ProyectoFinal/internal/handler"
+	"ProyectoFinal/internal/handler/inbound_order"
 	employeeRepository "ProyectoFinal/internal/repository/employee"
 	inboundOrderRepository "ProyectoFinal/internal/repository/inbound_order"
 	inboundOrderService "ProyectoFinal/internal/service/inbound_order"
 	"database/sql"
 )
 
-func GetInboundOrderHandler(db *sql.DB) *handler.InboundOrderHandler {
+func GetInboundOrderHandler(db *sql.DB) *inbound_order.InboundOrderHandler {
 	inboundOrderRepo := inboundOrderRepository.NewMySQLRepository(db)
 	employeeRepo := employeeRepository.NewMySQLRepository(db)
 	inboundOrderSrv := inboundOrderService.NewService(inboundOrderRepo, employeeRepo)
-	inboundOrderHdl := handler.NewInboundOrderHandler(inboundOrderSrv)
+	inboundOrderHdl := inbound_order.NewInboundOrderHandler(inboundOrderSrv)
 	return inboundOrderHdl
 }
