+++ internal/application/router/employee.go
@@ -1,13 +1,13 @@
 package router
 
 import (
-	"ProyectoFinal/internal/handler"
 	employeeHandler "ProyectoFinal/internal/handler/employee"
+	"ProyectoFinal/internal/handler/inbound_order"
 
 	"github.com/go-chi/chi/v5"
 )
 
-func EmployeeRoutes(ctr *employeeHandler.EmployeeHandler, inboundOrderHandler *handler.InboundOrderHandler) chi.Router {
+func EmployeeRoutes(ctr *employeeHandler.EmployeeHandler, inboundOrderHandler *inbound_order.InboundOrderHandler) chi.Router {
 	r := chi.NewRouter()
 	r.Get("/", ctr.GetAll())
 	r.Get("/{id}", ctr.GetById())
