+++ internal/application/router/inbound_order_router.go
@@ -1,12 +1,12 @@
 package router
 
 import (
-	"ProyectoFinal/internal/handler"
+	"ProyectoFinal/internal/handler/inbound_order"
 
 	"github.com/go-chi/chi/v5"
 )
 
-func InboundOrderRoutes(ctr *handler.InboundOrderHandler) chi.Router {
+func InboundOrderRoutes(ctr *inbound_order.InboundOrderHandler) chi.Router {
 	r := chi.NewRouter()
 	r.Post("/", ctr.Create())
 	return r
